Added files via upload
This is not finished, but should give some idea what it does. The only part unfinished is the control data field. We need to decide how we are going to represent the value of that control line (by string or bunch of 0,1). I will wait to see Dr. Szejda on this....